+++ packages/EasyCore/src/Bridge/Symfony/Validator/Constraints/Integer.php
@@ -7,6 +7,8 @@ namespace EonX\EasyCore\Bridge\Symfony\Validator\Constraints;
 use Symfony\Component\Validator\Constraint;
 
 /**
+ * @deprecated Use \EonX\EasyUtils\Validator\Constraints\Integer instead.
+ *
  * @Annotation
  * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
  */
