+++ packages/EasyDoctrine/composer.json
@@ -1,12 +1,15 @@
 {
     "name": "eonx-com/easy-doctrine",
     "description": "Provides functionalities to enhance Doctrine integration in your project",
-    "keywords": ["doctrine"],
+    "keywords": [
+        "doctrine"
+    ],
     "type": "symfony-bundle",
     "license": "MIT",
     "require": {
         "php": "^8.1",
         "eonx-com/easy-event-dispatcher": "^4.4",
+        "eonx-com/easy-utils": "^4.4",
         "nesbot/carbon": "^2.22",
         "myclabs/deep-copy": "^1.11"
     },
