+++ packages/EasyUtils/src/Validator/Constraints/AbstractNumberComparisonValidator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace EonX\EasyUtils\Validator\Constraints;
 
-use EonX\EasyUtils\ValueObject\Number as NumberValueObject;
+use EonX\EasyUtils\ValueObjects\Number as NumberValueObject;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\AbstractComparisonValidator;
 
@@ -22,7 +22,7 @@ abstract class AbstractNumberComparisonValidator extends AbstractComparisonValid
     }
 
     /**
-     * @param \EonX\EasyUtils\ValueObject\Number|int|string $value
+     * @param \EonX\EasyUtils\ValueObjects\Number|int|string $value
      */
     protected function formatValue(mixed $value, ?int $format = null): string
     {
